Разобьём утилитарный модуль на несколько модулей: common, task и render. Метод render и renderTemplate научим вызывать getElement у наших компонентов, чтобы код в месте использования стал чище. Заведём метод replace - обёртку над replaceChild. Заведём метод remove - обёртку для вызова методов удаления.
